orienting synthetics workflow on experiment
