+++ python/proteolizardalgo/hardware_models.py
@@ -1,0 +1,107 @@
+from abc import ABC, abstractmethod
+
+import tensorflow as tf
+import numpy as np
+import pandas as pd
+
+from proteolizardalgo.chemistry import  ccs_to_one_over_reduced_mobility
+
+
+class LiquidChromatographyApexModel(ABC):
+    def __init__(self):
+        pass
+
+class LiquidChromatographyProfileModel(ABC):
+    def __init__(self):
+        pass
+
+
+class NeuralChromatographyApex(LiquidChromatographyApexModel):
+
+    def __init__(self, model_path: str, tokenizer: tf.keras.preprocessing.text.Tokenizer):
+        super().__init__()
+        self.model = tf.keras.models.load_model(model_path)
+        self.tokenizer = tokenizer
+
+    def sequences_to_tokens(self, sequences: np.array) -> np.array:
+        print('tokenizing sequences...')
+        seq_lists = [list(s) for s in sequences]
+        tokens = self.tokenizer.texts_to_sequences(seq_lists)
+        tokens_padded = tf.keras.preprocessing.sequence.pad_sequences(tokens, 50, padding='post')
+        return tokens_padded
+
+    def sequences_tf_dataset(self, sequences: np.array, batched: bool = True, bs: int = 2048) -> tf.data.Dataset:
+        tokens = self.sequences_to_tokens(sequences)
+        print('generating tf dataset...')
+        pseudo_target = np.expand_dims(np.zeros_like(tokens[:, 0]), axis=1)
+
+        if batched:
+            return tf.data.Dataset.from_tensor_slices((tokens, pseudo_target)).batch(bs)
+        return tf.data.Dataset.from_tensor_slices((tokens, pseudo_target))
+
+    def get_retention_times(self, data: pd.DataFrame) -> np.array:
+        ds = self.sequences_tf_dataset(data['sequence-tokenized'])
+        print('predicting irts...')
+        return self.model.predict(ds)
+
+
+class IonMobilityApexModel(ABC):
+    def __init__(self):
+        pass
+
+class IonMobilityProfileModel(ABC):
+    def __init__(self):
+        pass
+
+class NeuralMobilityApex(IonMobilityApexModel):
+
+    def __init__(self, model_path: str, tokenizer: tf.keras.preprocessing.text.Tokenizer):
+        super().__init__()
+        self.model = tf.keras.models.load_model(model_path)
+        self.tokenizer = tokenizer
+
+    def sequences_to_tokens(self, sequences: np.array) -> np.array:
+        print('tokenizing sequences...')
+        seq_lists = [list(s) for s in sequences]
+        tokens = self.tokenizer.texts_to_sequences(seq_lists)
+        tokens_padded = tf.keras.preprocessing.sequence.pad_sequences(tokens, 50, padding='post')
+        return tokens_padded
+
+    def sequences_tf_dataset(self, mz: np.array, charges: np.array, sequences: np.array,
+                             batched: bool = True, bs: int = 2048) -> tf.data.Dataset:
+        tokens = self.sequences_to_tokens(sequences)
+        mz = np.expand_dims(mz, 1)
+        c = tf.one_hot(charges - 1, depth=4)
+        print('generating tf dataset...')
+        pseudo_target = np.expand_dims(np.zeros_like(tokens[:, 0]), axis=1)
+
+        if batched:
+            return tf.data.Dataset.from_tensor_slices(((mz, c, tokens), pseudo_target)).batch(bs)
+        return tf.data.Dataset.from_tensor_slices(((mz, c, tokens), pseudo_target))
+
+    def get_mobilities_and_ccs(self, data: pd.DataFrame) -> np.array:
+        ds = self.sequences_tf_dataset(data['mz'], data['charge'], data['sequence-tokenized'])
+
+        mz = data['mz'].values
+
+        print('predicting mobilities...')
+        ccs, _ = self.model.predict(ds)
+        one_over_k0 = ccs_to_one_over_reduced_mobility(np.squeeze(ccs), mz, data['charge'].values)
+
+        return np.c_[ccs, one_over_k0]
+
+class IonizationModel(ABC):
+    def __init__(self):
+        pass
+
+    @abstractmethod
+    def ionize():
+        pass
+
+class RandomIonSource(IonizationModel):
+    def __init__(self):
+        super().__init__()
+
+    def ionize(self, data, allowed_charges: list = [1, 2, 3, 4], p: list = [0.1, 0.5, 0.3, 0.1]):
+        return np.random.choice(allowed_charges, data.shape[0], p=p)
+
